requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
psutil==5.9.8
GitPython==3.1.41 